move v1 profiler controller to core utils
